requirements.txt
django
djangorestframework==3.14.0
django-filter==23.2